Code/data_collection/ky038_cough_model/requirements.txt
numpy
scipy
librosa
scikit-learn
joblib
pyserial
requests
soundfile
fastapi
uvicorn
python-multipart
